Group the Param views into Detail, List and Provider folders
Split ParamBadge.swift into ParamIcon and ParamBadge and move the kind colors next to the other ParamValue display helpers.
